+++ Sources/RyzenStatus/Core/FeaturePresets.swift
@@ -76,9 +76,10 @@ extension AppFeature {
     var energyProfile: FeatureEnergyProfile {
         switch self {
         case .scrollInverter, .smoothScroll, .windowMaximizer, .middleClick,
-             .mouseNavigation, .dockPreview, .dockClick, .shelf:
+             .mouseNavigation, .dockPreview, .dockClick, .shelf,
+             .mouseButtonShortcuts:
             return .mouse
-        case .switcher, .keyboardDebounce, .finderCutPaste:
+        case .switcher, .keyboardDebounce, .finderCutPaste, .superKey:
             return .keyboard
         case .textSnippets, .autoQuit:
             return .inputs